pyproject.toml
[project]
name = "cnam-intelligence-artificielle"
version = "1.0.0"
requires-python = "==3.14.3"
dependencies = ["pillow==12.1.1"]

[dependency-groups]
dev = ["pytest==9.0.2", "pytest-cov==7.0.0", "ruff==0.15.1", "ty==0.0.17"]

[tool.uv]
add-bounds = "exact"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "CPY",
    "ERA",
    "T20",
    "EM101",
    "PLR2004",
    "TRY003",
    "S101",
    "COM812",
]
